backend/requirements.txt
fastapi~=0.115.8
starlette~=0.45.3
pydantic~=2.10.6
python-dotenv~=1.0.1
sqlmodel~=0.0.22
uvicorn~=0.34.0
PyJWT~=2.10.1
passlib~=1.7.4
SQLAlchemy~=2.0.38
pydantic[email]
psycopg2-binary
bcrypt
python-multipart